Minor changes and bugfixes
- Add a label with last content status update time in language table
- Add limit to number of entries in the history database tables
- Databases are now closed on program exit to delete temporary files
- BUGFIX: databases: are now reloaded after changing the data path
- BUGFIX: kanji-section: remove nodes before recreating after update
- BUGFIX: reviews-timeline: don't repeat when running out of data now

+++ js/lib/data-managers/history.js
@@ -40,7 +40,20 @@ module.exports = function (paths, modules) {
 
     history.load = async function (language) {
         const path = paths.languageData(language).history;
-        dataMap[language] = promisifyDatabase(new sqlite3.Database(path));
+        const db = new sqlite3.Database(path);
+        dataMap[language] = promisifyDatabase(db);
+        dataMap[language].close = (callback) => db.close(callback);
+        const dbc = promisifyDatabase(db);
+        // Cache sizes of the tables
+        dataMap[language].sizes = { };
+        dataMap[language].sizes["dictionary"] =
+            (await dbc.all("SELECT COUNT(*) AS size FROM dictionary"))[0].size;
+        if (language === "Japanese") {
+            dataMap[language].sizes["kanji_info"] =
+                (await dbc.all("SELECT COUNT(*) AS c FROM kanji_info"))[0].c;
+            dataMap[language].sizes["kanji_search"] =
+                (await dbc.all("SELECT COUNT(*) AS c FROM kanji_search"))[0].c;
+        }
         await dataMap[language].run("PRAGMA wal_autocheckpoint=50");
         await dataMap[language].run("BEGIN TRANSACTION");
     };
@@ -59,6 +72,18 @@ module.exports = function (paths, modules) {
         data = dataMap[language];
     };
 
+    history.close = async function (language) {
+        return new Promise((resolve, reject) => {
+            dataMap[language].close((error) => error ? reject(error):resolve());
+        });
+    };
+
+    history.closeAll = async function () {
+        for (const language in dataMap) {
+            await history.close(language);
+        }
+    };
+
     history.addEntry = async function (table, entry) {
         entry.time = utility.getTime();
         const keys = [];
@@ -69,10 +94,33 @@ module.exports = function (paths, modules) {
             qMarks.push("?");
             values.push(entry[key]);
         }
-        // Delete old entry with same name if there is one
-        await data.run(`DELETE FROM ${table} WHERE name = ?`, entry.name);
+        // Delete old entry with the same name if there is one
+        const entryExists = (await data.all(
+            `SELECT COUNT(*) AS count FROM ${table} WHERE name = ?`,
+            entry.name))[0].count;
+        if (entryExists) {
+            await data.run(`DELETE FROM ${table} WHERE name = ?`, entry.name);
+            --data.sizes[table];
+        }
+        // Add the entry and increase the cached size counter
         await data.run(`INSERT INTO ${table} (${keys.join(", ")})
                         VALUES (${qMarks.join(", ")})`, ...values);
+        ++data.sizes[table];
+        // Get maximum allowed number of database entries from the settings
+        let limit;
+        if (table === "dictionary") {
+            limit = dataManager.settings.dictionary.historySizeLimit;
+        } else if (table === "kanji_search") {
+            limit = dataManager.settings.kanjiSearch.historySizeLimit;
+        } else if (table === "kanji_info") {
+            limit = dataManager.settings.kanjiInfo.historySizeLimit;
+        }
+        // If the database size is above limit, delete the least recent entry
+        if (data.sizes[table] > limit) {
+            const lastEntryId = (await data.all(
+                `SELECT id FROM ${table} ORDER BY id ASC LIMIT 1`))[0].id;
+            await data.run(`DELETE FROM ${table} WHERE id = ?`, lastEntryId);
+        }
     };
 
     history.get = async function (table) {
